refactor(home): Relocate home screen to features directory
This commit refactors the file structure by moving all home-related presentation layer files from `lib/presentation/home` to a new `lib/presentation/features/home` directory.

This change improves code organization by grouping files by feature. All relevant imports have been updated accordingly across the application to reflect the new file paths.
