+++ lib/presentation/features/home/home_screen.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:sellio_mobile/core/design_system/themes/sellio_theme.dart';
-import '../screens/store_details/store_details_screen.dart';
+import '../../screens/store_details/store_details_screen.dart';
 import 'cubit/home_cubit.dart';
 import 'cubit/home_state.dart';
 import 'widgets/home_app_bar.dart';
