+++ lib/presentation/features/home/widgets/top_stores_section.dart
@@ -1,9 +1,10 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_svg/svg.dart';
 import 'package:sellio_mobile/core/design_system/constants/assets.dart';
-import '../../../core/design_system/widgets/cards/store_card.dart';
-import '../../../core/design_system/widgets/section_header.dart';
-import '../../../domain/entities/store.dart';
+
+import '../../../../core/design_system/widgets/cards/store_card.dart';
+import '../../../../core/design_system/widgets/section_header.dart';
+import '../../../../domain/entities/store.dart';
 
 class TopStoresSection extends StatelessWidget {
   final List<Store> topStores;
